requirements.txt
Django==5.0.4
django_allauth==0.62.1
